separated out createCORSRequest to its own file
