+++ LibraryAccess/src/components/SearchResults.js
@@ -1,26 +1,10 @@
 import React from 'react';
 
 import {CardView} from 'components/CardView';
-import {apiURL} from './Home'
+import {apiURL, createCORSRequest} from 'sources/createCORSRequest';
 import {PageHeader} from "components/PageHeader";
 import {PageFooter} from "components/PageFooter";
 import {Link} from "react-router-dom";
-
-function createCORSRequest(method, url) {
-  var xhr = new XMLHttpRequest();
-  if ("withCredentials" in xhr) {
-    // XHR for Chrome/Firefox/Opera/Safari.
-    xhr.open(method, url, false);
-  } else if (typeof XDomainRequest != "undefined") {
-    // XDomainRequest for IE.
-    xhr = new XDomainRequest();
-    xhr.open(method, url);
-  } else {
-    // CORS not supported.
-    xhr = null;
-  }
-  return xhr;
-}
 
 class SearchResults extends React.Component {
   constructor(props) {
