+++ LibraryAccess/src/sources/createCORSRequest.js
@@ -1,0 +1,20 @@
+function createCORSRequest(method, url) {
+  var xhr = new XMLHttpRequest();
+  if ('withCredentials' in xhr) {
+    // XHR for Chrome/Firefox/Opera/Safari.
+    xhr.open(method, url, false);
+  } else if (typeof XDomainRequest != 'undefined') {
+    // XDomainRequest for IE.
+    xhr = new XDomainRequest();
+    xhr.open(method, url);
+  } else {
+    // CORS not supported.
+    xhr = null;
+  }
+  return xhr;
+}
+
+
+const apiURL = 'http://142.93.181.3:7001';
+
+export { apiURL, createCORSRequest };
